requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
validators==0.34.0
tzdata==2025.1
validators==0.34.0
